requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
attrs==25.3.0
billiard==4.2.2
bleach==6.2.0
boto3==1.40.40
botocore==1.40.40
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
Django==5.2.6
django-cors-headers==4.9.0
django-csp==4.0
django-environ==0.12.0
django-filter==25.1
django-guardian==3.2.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.11
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
numpy==2.3.3
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
sentry-sdk==2.39.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
gunicorn==23.0.0
whitenoise==6.6.0